multi-agent-region-expansion-planner/requirements.txt
strands-agents-tools
mcp[cli]
boto3
botocore